pyproject.toml
[tool.poetry]
name = "bghinsights"
version = "0.1.0"
description = ""
authors = ["cspr00 <caspar.galen@progius.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
adlfs = "^2024.4.1"
python-dotenv = "^1.0.1"
pdfminer-six = "^20231228"
requests= "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
